Support multiple ratings
[#166438585]

+++ video-service/src/main/kotlin/com/boclips/videos/service/presentation/video/VideoResource.kt
@@ -32,7 +32,7 @@ data class VideoResource(
     @get:JsonView(PublicApiProjection::class)
     val ageRange: AgeRangeResource? = null,
     @get:JsonView(PublicApiProjection::class)
-    val rating: Int? = null,
+    val rating: Double? = null,
     @get:JsonView(PublicApiProjection::class)
     val source: String? = null,
